+++ src/ckanapi_harvesters/policies/__init__.py
@@ -8,6 +8,7 @@ POLICY_FILE_FORMAT_VERSION = "0.0.0"  # version of the data format policy file f
 
 from . import data_format_policy_defs
 from . import data_format_policy_errors
+from . import policy_report
 from . import data_format_policy_abc
 from . import data_format_policy_lists
 from . import data_format_policy_tag_groups
@@ -16,6 +17,7 @@ from . import data_format_policy_output_config
 from . import data_format_policy
 
 # usage shortcuts
+from ckanapi_harvesters.policies.policy_report import PackagePolicyReport
 from ckanapi_harvesters.policies.data_format_policy import CkanPackageDataFormatPolicy